+++ projects/core/src/ae2Integration/kotlin/site/siredvin/peripheralworks/integrations/ae2/AE2Helper.kt
@@ -1,3 +1,5 @@
+@file:Suppress("DEPRECATION")
+
 package site.siredvin.peripheralworks.integrations.ae2
 
 import appeng.api.stacks.*