+++ projects/core/src/ae2Integration/kotlin/site/siredvin/peripheralworks/integrations/ae2/MENetworkBlockPlugin.kt
@@ -5,7 +5,6 @@ import appeng.api.networking.security.IActionSource
 import appeng.api.stacks.AEFluidKey
 import appeng.api.stacks.AEItemKey
 import appeng.api.stacks.AEKey
-import appeng.blockentity.grid.AENetworkBlockEntity
 import dan200.computercraft.api.lua.LuaFunction
 import dan200.computercraft.api.lua.MethodResult
 import net.minecraft.core.BlockPos
@@ -22,7 +21,7 @@ import site.siredvin.tweakium.modules.peripheral.representation.LuaRepresentatio
 import java.util.*
 import kotlin.NoSuchElementException
 
-class MENetworkBlockPlugin(private val level: Level, private val entity: AENetworkBlockEntity) : IPeripheralPlugin {
+class MENetworkBlockPlugin(private val level: Level, private val entity: MENetworkPeripheralBlockEntity) : IPeripheralPlugin {
     companion object {
         const val PLUGIN_TYPE = "ae2"
     }
@@ -36,34 +35,37 @@ class MENetworkBlockPlugin(private val level: Level, private val entity: AENetwo
                 return null
             }
             val entity = level.getBlockEntity(pos)
-            if (entity !is AENetworkBlockEntity) {
+            if (entity !is MENetworkPeripheralBlockEntity) {
                 return null
             }
             return MENetworkBlockPlugin(level, entity)
         }
     }
 
+    private val grid
+        get() = entity.mainNode.takeIf { it.isActive }?.grid
+
     @LuaFunction(mainThread = true)
     fun getAverageEnergyDemand(): Double {
-        val energyService = entity.mainNode.grid?.energyService ?: return 0.0
+        val energyService = grid?.energyService ?: return 0.0
         return energyService.avgPowerUsage
     }
 
     @LuaFunction(mainThread = true)
     fun getAverageEnergyIncome(): Double {
-        val energyService = entity.mainNode.grid?.energyService ?: return 0.0
+        val energyService = grid?.energyService ?: return 0.0
         return energyService.avgPowerInjection
     }
 
     @LuaFunction(mainThread = true)
     fun getChannelEnergyDemand(): Double {
-        val energyService = entity.mainNode.grid?.energyService ?: return 0.0
+        val energyService = grid?.energyService ?: return 0.0
         return energyService.channelPowerUsage
     }
 
     @LuaFunction(mainThread = true)
     fun getChannelInformation(): Map<String, Any> {
-        val pathingService = entity.mainNode.grid?.pathingService ?: return emptyMap()
+        val pathingService = grid?.pathingService ?: return emptyMap()
         return mapOf(
             "maxChannels" to pathingService.channelMode.adHocNetworkChannels,
             "usedChannels" to pathingService.usedChannels,
@@ -72,7 +74,7 @@ class MENetworkBlockPlugin(private val level: Level, private val entity: AENetwo
 
     @LuaFunction(mainThread = true)
     fun getCraftingCPUs(): List<Map<String, *>> {
-        val craftingService = entity.mainNode.grid?.craftingService ?: return emptyList()
+        val craftingService = grid?.craftingService ?: return emptyList()
         val data: MutableList<Map<String, *>> = mutableListOf()
         craftingService.cpus.forEach {
             val cpuInformation = mutableMapOf<String, Any>()
@@ -89,7 +91,7 @@ class MENetworkBlockPlugin(private val level: Level, private val entity: AENetwo
 
     @LuaFunction(mainThread = true)
     fun getCraftableItems(): List<Map<String, *>> {
-        val craftingService = entity.mainNode.grid?.craftingService ?: return emptyList()
+        val craftingService = grid?.craftingService ?: return emptyList()
         val data: MutableList<Map<String, *>> = mutableListOf()
         craftingService.getCraftables { it is AEItemKey }.forEach {
             data.add(LuaRepresentation.forItem((it as AEItemKey).item))
@@ -99,7 +101,7 @@ class MENetworkBlockPlugin(private val level: Level, private val entity: AENetwo
 
     @LuaFunction(mainThread = true)
     fun getCraftableFluids(): List<Map<String, *>> {
-        val craftingService = entity.mainNode.grid?.craftingService ?: return emptyList()
+        val craftingService = grid?.craftingService ?: return emptyList()
         val data: MutableList<Map<String, *>> = mutableListOf()
         craftingService.getCraftables { it is AEFluidKey }.forEach {
             data.add(
@@ -113,7 +115,7 @@ class MENetworkBlockPlugin(private val level: Level, private val entity: AENetwo
 
     @LuaFunction(mainThread = true)
     fun getPatternsFor(mode: String, id_key: String): List<Map<String, *>> {
-        val craftingService = entity.mainNode.grid?.craftingService ?: return emptyList()
+        val craftingService = grid?.craftingService ?: return emptyList()
         val key: AEKey = buildKey(mode, id_key)
         val patterns = craftingService.getCraftingFor(key)
         val data = mutableListOf<Map<String, *>>()
@@ -149,7 +151,7 @@ class MENetworkBlockPlugin(private val level: Level, private val entity: AENetwo
 
     @LuaFunction(mainThread = true)
     fun getActiveCraftings(): MethodResult {
-        val craftingService = entity.mainNode.grid?.craftingService ?: return MethodResult.of(null, "AE2 network is not connected")
+        val craftingService = grid?.craftingService ?: return MethodResult.of(null, "AE2 network is not connected")
         val craftingList: MutableList<Map<String, Any>> = mutableListOf()
         craftingService.cpus.forEach {
             if (it.isBusy && it.jobStatus != null) {
@@ -170,7 +172,7 @@ class MENetworkBlockPlugin(private val level: Level, private val entity: AENetwo
 
     @LuaFunction(mainThread = false)
     fun scheduleCrafting(mode: String, id_key: String, amount: Optional<Long>, targetCPU: Optional<String>): MethodResult {
-        val craftingService = entity.mainNode.grid?.craftingService ?: return MethodResult.of(null, "AE2 network is not connected")
+        val craftingService = grid?.craftingService ?: return MethodResult.of(null, "AE2 network is not connected")
 
         val key = buildKey(mode, id_key)
         val source = IActionSource.ofMachine(entity)